+++ packages/react-ui/components/MenuItem/MenuItem.styles.ts
@@ -132,7 +132,7 @@ export const getStyles = memoizeGetStyles(({ css }: Emotion) => ({
   commentHover(t: Theme) {
     return css`
       color: ${t.menuItemCommentColorHover};
-      opacity: 0.6;
+      opacity: ${t.menuItemCommentOpacityHover};
     `;
   },
   icon() {